Refactor VenueDetailPage component
