+++ src/pages/venue/VenueDetailPage.tsx
@@ -1,31 +1,46 @@
-// src/pages/venue/VenueDetailPage.tsx
+/**
+ * @file VenueDetailPage.tsx
+ * @description Page for viewing detailed information about a venue
+ */
+
 import { useState, useEffect } from 'react';
 import { useParams, Link, useNavigate, useLocation } from 'react-router-dom';
 import { getVenueById } from '../../api/venueService';
 import { createBooking } from '../../api/bookingService';
-import { getVenueRatings, hasUserRatedVenue, calculateAverageRating, getVenueRatingInfo } from '../../api/ratingService';
+import { getVenueRatings, hasUserRatedVenue, calculateAverageRating } from '../../api/ratingService';
 import { Rating } from '../../types/rating';
 import { Venue } from '../../types/venue';
 import { useAuth } from '../../contexts/AuthContext';
+import { addToRecentlyViewed } from '../../utils/venueUtils';
+
+// Components
 import ImageGallery from '../../components/venue/ImageGallery';
-import AmenityIcon from '../../components/venue/AmenityIcon';
 import BookingCalendar from '../../components/venue/BookingCalendar';
-import RatingForm from '../../components/venue/RatingForm';
-import RatingsList from '../../components/venue/RatingsList';
 import RatingPrompt from '../../components/venue/RatingPrompt';
 import RatingModal from '../../components/venue/RatingModal';
-import StarRating from '../../components/common/StarRating';
-import { getUserAvatar } from '../../utils/avatarUtils';
+import VenueHeader from '../../components/venue/VenueHeader';
+import VenueAmenities from '../../components/venue/VenueAmenities';
+import VenueLocation from '../../components/venue/VenueLocation';
+import VenueHost from '../../components/venue/VenueHost';
+import VenueRatingSection from '../../components/venue/VenueRatingSection';
+import VenuePriceCard from '../../components/venue/VenuePriceCard';
 
+/**
+ * Page component for viewing detailed venue information and making bookings
+ * 
+ * @returns {JSX.Element} Rendered component
+ */
 const VenueDetailPage = () => {
   const { id } = useParams<{ id: string }>();
   const { token, user } = useAuth();
   const navigate = useNavigate();
 
+  // Get source from query parameters
   const currentLocation = useLocation();
   const queryParams = new URLSearchParams(currentLocation.search);
   const source = queryParams.get('source');
   
+  // Page state
   const [venue, setVenue] = useState<Venue | null>(null);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
@@ -41,7 +56,9 @@ const VenueDetailPage = () => {
   const [showRatingForm, setShowRatingForm] = useState(false);
   const [showRatingModal, setShowRatingModal] = useState(false);
 
-  // Function to fetch ratings from localStorage
+  /**
+   * Fetches ratings data for the venue
+   */
   const fetchRatings = async () => {
     if (!id) return;
     
@@ -58,20 +75,20 @@ const VenueDetailPage = () => {
       
       // Check if current user has already rated
       if (user) {
-        // Use the imported hasUserRatedVenue directly with the venue ID and user name
+        // Check if user has rated using the imported utility
         const userHasRated = hasUserRatedVenue(id, user.name);
         setUserHasRated(userHasRated);
       }
     } catch (err) {
-      console.error('Error fetching ratings:', err);
+      // Handle error silently
     } finally {
       setRatingsLoading(false);
     }
   };
 
-  // Check if the user has booked this venue
-  // For development purposes, consider everyone eligible to rate
-  // In a real app, you'd check if user has completed a booking
+  /**
+   * Checks if the user is eligible to rate this venue
+   */
   const checkEligibilityToRate = () => {
     // In development, allow all authenticated users to rate
     // For production, uncomment the code below to check booking history
@@ -82,12 +99,15 @@ const VenueDetailPage = () => {
         new Date(booking.dateTo) < new Date()  // Only past bookings
       );
       setUserHasBooked(userCompletedBookings.length > 0);
-  } else {
-    // For now, allow all authenticated users to rate
-    setUserHasBooked(false);
-  }
-};
+    } else {
+      // For now, allow all authenticated users to rate
+      setUserHasBooked(false);
+    }
+  };
 
+  /**
+   * Fetches venue data and manages related state
+   */
   useEffect(() => {
     const fetchVenue = async () => {
       if (!id) return;
@@ -103,9 +123,10 @@ const VenueDetailPage = () => {
         checkEligibilityToRate();
         
         // Add this venue to recently viewed (localStorage)
-        addToRecentlyViewed(venueData);
+        if (venueData) {
+          addToRecentlyViewed(venueData.id, venueData.name);
+        }
       } catch (err) {
-        console.error('Error fetching venue:', err);
         setError('Failed to load venue details. Please try again later.');
       } finally {
         setLoading(false);
@@ -116,54 +137,21 @@ const VenueDetailPage = () => {
     fetchRatings();
   }, [id, user]);
 
-  // Add venue to recently viewed in localStorage
-  const addToRecentlyViewed = (venue: Venue) => {
-    try {
-      console.log("Adding to recently viewed:", venue.id, venue.name);
-      
-      // Get existing recently viewed IDs or initialize empty array
-      const recentlyViewedString = localStorage.getItem('recentlyViewed');
-      let recentlyViewedIds: string[] = [];
-      
-      if (recentlyViewedString) {
-        try {
-          const parsed = JSON.parse(recentlyViewedString);
-          // Check if it's an array of strings (IDs) or an array of objects (old format)
-          if (Array.isArray(parsed)) {
-            if (parsed.length > 0 && typeof parsed[0] === 'string') {
-              // It's already the correct format - array of IDs
-              recentlyViewedIds = parsed;
-            } else if (parsed.length > 0 && typeof parsed[0] === 'object') {
-              // Old format - array of venue objects
-              recentlyViewedIds = parsed.map((v: any) => v.id || '').filter(Boolean);
-            }
-          }
-        } catch (error) {
-          console.error('Error parsing recently viewed:', error);
-        }
-      }
-      
-      // Remove this venue ID if it already exists
-      recentlyViewedIds = recentlyViewedIds.filter(id => id !== venue.id);
-      
-      // Add to beginning of array
-      recentlyViewedIds.unshift(venue.id);
-      
-      // Keep only the last 5
-      const limited = recentlyViewedIds.slice(0, 5);
-      
-      localStorage.setItem('recentlyViewed', JSON.stringify(limited));
-    } catch (error) {
-      console.error('Error saving to recently viewed:', error);
-    }
-  };
-
-  // Handle rating submission - simplified for localStorage version
+  /**
+   * Handles rating submission completion
+   */
   const handleRatingSubmitted = () => {
     // Refresh ratings after submission
     fetchRatings();
   };
 
+  /**
+   * Handles booking submission
+   * 
+   * @param {Date} from - Check-in date
+   * @param {Date} to - Check-out date
+   * @param {number} guests - Number of guests
+   */
   const handleBookingSubmit = async (from: Date, to: Date, guests: number) => {
     if (!id || !token) return;
     
@@ -190,11 +178,11 @@ const VenueDetailPage = () => {
       // Scroll to the top to show the success message
       window.scrollTo(0, 0);
     } catch (err: any) {
-      console.error('Error creating booking:', err);
       setBookingError('Failed to create booking. Please try again.');
     }
   };
 
+  // Loading state
   if (loading) {
     return (
       <div className="flex justify-center items-center min-h-screen">
@@ -203,6 +191,7 @@ const VenueDetailPage = () => {
     );
   }
 
+  // Error state
   if (error || !venue) {
     return (
       <div className="max-w-6xl mx-auto px-4 py-8">
@@ -271,27 +260,12 @@ const VenueDetailPage = () => {
       </Link>
 
       {/* Venue Header */}
-      <div className="mb-8">
-        <h1 className="text-3xl md:text-4xl font-bold mb-2 font-averia">{name}</h1>
-        <div className="flex flex-wrap items-center text-gray-600 gap-4">
-          <div className="flex items-center">
-            <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5 mr-1" viewBox="0 0 20 20" fill="currentColor">
-              <path fillRule="evenodd" d="M5.05 4.05a7 7 0 119.9 9.9L10 18.9l-4.95-4.95a7 7 0 010-9.9zM10 11a2 2 0 100-4 2 2 0 000 4z" clipRule="evenodd" />
-            </svg>
-            <span>
-              {[location.city, location.country].filter(Boolean).join(', ') || 'Location not specified'}
-            </span>
-          </div>
-          {averageRating > 0 && (
-            <div className="flex items-center">
-              <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5 text-yellow-500 mr-1" viewBox="0 0 20 20" fill="currentColor">
-                <path d="M9.049 2.927c.3-.921 1.603-.921 1.902 0l1.07 3.292a1 1 0 00.95.69h3.462c.969 0 1.371 1.24.588 1.81l-2.8 2.034a1 1 0 00-.364 1.118l1.07 3.292c.3.921-.755 1.688-1.54 1.118l-2.8-2.034a1 1 0 00-1.175 0l-2.8 2.034c-.784.57-1.838-.197-1.539-1.118l1.07-3.292a1 1 0 00-.364-1.118L2.98 8.72c-.783-.57-.38-1.81.588-1.81h3.461a1 1 0 00.951-.69l1.07-3.292z" />
-              </svg>
-              <span>{averageRating.toFixed(1)} ({ratings.length} {ratings.length === 1 ? 'review' : 'reviews'})</span>
-            </div>
-          )}
-        </div>
-      </div>
+      <VenueHeader 
+        name={name}
+        location={location}
+        averageRating={averageRating}
+        ratingsCount={ratings.length}
+      />
 
       {/* Image Gallery */}
       <div className="mb-6">
@@ -319,165 +293,35 @@ const VenueDetailPage = () => {
           </section>
 
           {/* Amenities Section */}
-          <section className="mb-10">
-            <h2 className="text-2xl font-bold mb-4 font-averia">Amenities</h2>
-            <div className="grid grid-cols-2 md:grid-cols-4 gap-4">
-              <div className={`flex items-center ${meta?.wifi ? 'text-gray-800' : 'text-gray-400'}`}>
-                <AmenityIcon type="wifi" active={meta?.wifi || false} />
-                <span className="ml-2">WiFi</span>
-              </div>
-              <div className={`flex items-center ${(meta?.parking ?? false) ? 'text-gray-800' : 'text-gray-400'}`}>
-                <AmenityIcon type="parking" active={meta?.parking ?? false} />
-                <span className="ml-2">Parking</span>
-              </div>
-              <div className={`flex items-center ${(meta?.breakfast ?? false) ? 'text-gray-800' : 'text-gray-400'}`}>
-                <AmenityIcon type="breakfast" active={meta?.breakfast ?? false} />
-                <span className="ml-2">Breakfast</span>
-              </div>
-              <div className={`flex items-center ${(meta?.pets ?? false) ? 'text-gray-800' : 'text-gray-400'}`}>
-                <AmenityIcon type="pets" active={meta?.pets ?? false} />
-                <span className="ml-2">Pets Allowed</span>
-              </div>
-            </div>
-          </section>
+          <VenueAmenities meta={meta} />
 
           {/* Location Section */}
-          <section className="mb-10">
-            <h2 className="text-2xl font-bold mb-4 font-averia">Location</h2>
-            <div className="bg-gray-100 p-4 rounded-lg">
-              <h3 className="font-medium mb-2 font-averia">Address</h3>
-              <p>
-                {location.address && <span className="block">{location.address}</span>}
-                {location.city && location.zip && <span className="block">{location.city}, {location.zip}</span>}
-                {location.country && <span className="block">{location.country}</span>}
-              </p>
-              
-              {/* If we had map integration, it would go here */}
-              <div className="h-64 bg-gray-200 mt-4 rounded flex items-center justify-center">
-                <p className="text-gray-500">Map view not available</p>
-              </div>
-            </div>
-          </section>
+          <VenueLocation location={location} />
 
           {/* Host Section */}
-          {owner && (
-            <section className="mb-10">
-              <h2 className="text-2xl font-bold mb-4 font-averia">
-                Hosted by{' '}
-                <Link to={`/profiles/${owner.name}`} className="text-[#0081A7] hover:underline">
-                  {owner.name}
-                </Link>
-              </h2>
-              <div className="flex items-center">
-                <div className="w-16 h-16 bg-gray-200 rounded-full overflow-hidden mr-4">
-                  {/* Check for avatar in localStorage first */}
-                  {owner.name && getUserAvatar(owner.name) ? (
-                    <img 
-                      src={getUserAvatar(owner.name)} 
-                      alt={`${owner.name}'s avatar`} 
-                      className="w-full h-full object-cover"
-                      onError={(e) => {
-                        const target = e.target as HTMLImageElement;
-                        target.src = 'https://placehold.co/200x200?text=Host';
-                      }}
-                    />
-                  ) : owner.avatar ? (
-                    <img 
-                      src={owner.avatar.url} 
-                      alt={owner.avatar.alt || `${owner.name}'s avatar`} 
-                      className="w-full h-full object-cover"
-                      onError={(e) => {
-                        const target = e.target as HTMLImageElement;
-                        target.src = 'https://placehold.co/200x200?text=Host';
-                      }}
-                    />
-                  ) : (
-                    <div className="w-full h-full flex items-center justify-center text-gray-400">
-                      <svg xmlns="http://www.w3.org/2000/svg" className="h-8 w-8" viewBox="0 0 20 20" fill="currentColor">
-                        <path fillRule="evenodd" d="M10 9a3 3 0 100-6 3 3 0 000 6zm-7 9a7 7 0 1114 0H3z" clipRule="evenodd" />
-                      </svg>
-                    </div>
-                  )}
-                </div>
-                <div>
-                  <h3 className="font-bold text-lg font-averia">
-                    <Link to={`/profiles/${owner.name}`} className="text-[#0081A7] hover:underline">
-                      {owner.name}
-                    </Link>
-                  </h3>
-                  {owner.bio && <p className="text-gray-600">{owner.bio}</p>}
-                </div>
-              </div>
-            </section>
-          )}
+          {owner && <VenueHost owner={owner} />}
 
           {/* Ratings Section */}
-          <section className="mb-10">
-            <div className="flex justify-between items-center mb-4">
-              <h2 className="text-2xl font-bold font-averia">
-                Reviews
-                {ratings.length > 0 && ` (${ratings.length})`}
-              </h2>
-              {averageRating > 0 && (
-                <div className="flex items-center gap-1">
-                  <div className="flex">
-                    {[1, 2, 3, 4, 5].map((star) => (
-                      <span key={star} className="text-xl">
-                        {star <= Math.round(averageRating) ? (
-                          <span className="text-yellow-500">★</span>
-                        ) : (
-                          <span className="text-gray-300">☆</span>
-                        )}
-                      </span>
-                    ))}
-                  </div>
-                  <span className="ml-1 font-medium">{averageRating.toFixed(1)}</span>
-                </div>
-              )}
-            </div>
-            
-            {/* Show rating form if showRatingForm is true */}
-            {(showRatingForm && canRateVenue) && (
-              <RatingForm 
-                venueId={id || ''} 
-                onRatingSubmitted={handleRatingSubmitted}
-                canRate={true}
-              />
-            )}
-            
-            {/* Display ratings list */}
-            <div className="mt-6">
-              <RatingsList 
-                ratings={ratings} 
-                isLoading={ratingsLoading} 
-              />
-            </div>
-          </section>
+          <VenueRatingSection
+            averageRating={averageRating}
+            ratings={ratings}
+            isLoading={ratingsLoading}
+            showRatingForm={showRatingForm}
+            canRateVenue={canRateVenue ?? false}
+            venueId={id || ''}
+            onRatingSubmitted={handleRatingSubmitted}
+          />
         </div>
 
         {/* Right Column - Booking Calendar */}
         <div className="lg:w-1/3">
           <div className="sticky top-6">
-            <div className="bg-white rounded-lg border border-gray-200 p-6 mb-6">
-              <div className="flex justify-between items-start mb-4">
-                <div>
-                  <span className="text-2xl font-bold">${price}</span>
-                  <span className="text-gray-600"> / night</span>
-                </div>
-                {averageRating > 0 && (
-                  <div className="flex items-center">
-                    <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5 text-yellow-500 mr-1" viewBox="0 0 20 20" fill="currentColor">
-                      <path d="M9.049 2.927c.3-.921 1.603-.921 1.902 0l1.07 3.292a1 1 0 00.95.69h3.462c.969 0 1.371 1.24.588 1.81l-2.8 2.034a1 1 0 00-.364 1.118l1.07 3.292c.3.921-.755 1.688-1.54 1.118l-2.8-2.034a1 1 0 00-1.175 0l-2.8 2.034c-.784.57-1.838-.197-1.539-1.118l1.07-3.292a1 1 0 00-.364-1.118L2.98 8.72c-.783-.57-.38-1.81.588-1.81h3.461a1 1 0 00.951-.69l1.07-3.292z" />
-                    </svg>
-                    <span>{averageRating.toFixed(1)}</span>
-                  </div>
-                )}
-              </div>
-
-              <div className="mb-4">
-                <p className="text-gray-600 mb-1">Max guests: {maxGuests}</p>
-              </div>
-            </div>
+            {/* Price Card */}
+            <VenuePriceCard
+              price={price}
+              maxGuests={maxGuests}
+              averageRating={averageRating}
+            />
 
             {/* Enhanced Booking Calendar Component */}
             <BookingCalendar
